backend/requirements.txt
Django==3.2.17
requests==2.28.2
Pillow==9.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-colorfield==0.8.0
pathlib==1.0.1
sorl-thumbnail==12.9.0
